fix #1035: moves contexts to contexts/ dir
