+++ src/app/components/Brand/index.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Brand from './index';
 
 import { shouldMatchSnapshot } from '../../helpers/tests/testHelpers';
-import { ServiceContextProvider } from '../ServiceContext';
+import { ServiceContextProvider } from '../../contexts/ServiceContext';
 
 describe('Brand', () => {
   shouldMatchSnapshot(
